Behavior should be the same across adapters (all specs pass)
Previously, the specs affected in this commit were
guarded to either run on sqlite,mysql or postgres.
Restricting the specs to run on these adapters only
is not necessary (nor correct imho). The validation
behavior should not change across adapters.

I think that previously this restriction was only
there to ensure that at least one (of those) adapters
was correctly setup. Since we now rely on the spec
adapter provided by dm-core/spec/setup, we can be
sure that the desired adapter is already set up.
Therefore we don't need to guard these specs anymore.

All specs pass now on sqlite,mysql and postgres