Avoid unnecessary active_support/extlib collision
Previously, the code made the assumption that
AS >= 2.3.5 was being used, as core_ext/object/blank
was introduced in that version.  In cases where an
earlier version is used (e.g. 2.3.3), the result is
a partial load of AS and then Extlib's overrides,
because the LoadError will trigger due to a missing
require.

When used with inside a Rails <2.3.5 app, chances
are AS has already been loaded completely anyway,
thus the LoadError consistently produces a "constant
redefined" warning regarding OrderedHash.