Only validate dirty properties.
This is an efficiency optimzation meant to prevent triggering the loading of
lazy-designated attributes that aren't dirty and thus arguably don't need
(re)validation.

This could contradict previous intent that object.save should guarantee *all*
data validity, regardless of whether or not the user mutated the data.  But I'd
argue the common case is that users only mean to validate that which they
created/changed; if integrity of existing data is an issue, this is not really
the problem that DataMapper and Validations are meant to solve.