Do not optimize away checking non-dirty attributes
There are a lot of situations where this is a bad
move and can lead to invalid data in the database.
The attached testcase is a simple example of this
problem, where a change to another field
invalidates the resource. DM will now happily save
this resource, even though it is quite clearly
invalid.

We also removed the check where valid? isn't
called at all if the model isn't dirty. It's not
safe to assume that the model is valid just
because there are no dirty attributes, the
validations of the model class may have changed
since the model was added to the DB. With the
current architecture there is no way of even
checking *if* such invalid models exist.