Use Range#begin and #end instead of #min and #max.
Range gets #min and #max from Enumerable, which means it iterates to reach the endpoint values. Iterating instead of just querying for them is inefficient. Moreover, Ruby can't seem to iterate it's way to Infinity or -Infinity, which are valid Range endpoints (and which #begin and #end have no problem with).